Split menu module
